% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.coef.q}
\alias{ggplot.coef.q}
\title{Coefficients over the M-Quantiles}
\usage{
\method{ggplot}{coef.q}(fit, coef = FALSE)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{coef}{numeric, coefficient to plot. Defaults to \code{FALSE}.}
}
\description{
This plot shows the coefficients over the m-quantiles.
This can show the changing relationship for changing q.
}
\examples{
library(mq1)

df <- simulate_data( n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                  measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2 * x1",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "coef.q", coef = 0)
}
