% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.fit.res}
\alias{ggplot.fit.res}
\title{Fitted vs. Residuals Plot}
\usage{
\method{ggplot}{fit.res}(fit, sqr.res = FALSE, add.qvals = FALSE)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{sqr.res}{boolean controlling if the plot should show the squared residuals instead.}

\item{add.qvals}{numeric, containing additional m-quantiles to be shown in plot.}
}
\description{
The fit.res plot creates a residuals over fitted value plot.
The smoothed residual indicates the mean residual conditional on the fitted values.
Additionally, observation which are downweighted are marked with a triangle.
}
\examples{
library(mq1)

df <- simulate_data( n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                  measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2 * x1",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "fit.res")

}
