% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.fitted.observed}
\alias{ggplot.fitted.observed}
\title{Fitted vs. Observed Values}
\usage{
\method{ggplot}{fitted.observed}(fit, add.qvals = FALSE)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{add.qvals}{numeric, containing additional m-quantiles to be shown in plot.}
}
\description{
This plot shows the fitted values over the observed. This gives a good indication of how good model fit is-
Additionally observation which are downweighted are marked with a triangle.
}
\examples{
library(mq1)

df <- simulate_data( n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2 * x1",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "fitted.observed")
}
