% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.optk}
\alias{ggplot.optk}
\title{Optimal k-Plot}
\usage{
\method{ggplot}{optk}(fit, kstart = 0.5, kend = 3, kstep = 0.1)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{kstart}{numeric, the first k estimated and shown in the plot.}

\item{kend}{numeric, the last k estimated and shown in the plot.}

\item{kstep}{numeric, controls the steps in which k is increased between kstart and kend.}
}
\description{
This plot shows the distance of the residual distribution to the normal distribution.
It there by uses the method proposed by James Dawber (2017). The optimal k is following
the method the k with the smallest distance, hence the minimum in the plot.
}
\examples{
library(mq1)

df <- simulate_data(n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2 * x1",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "optk")
}
