% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.prop.huber.k}
\alias{ggplot.prop.huber.k}
\title{Proportion of residuals huberised over k}
\usage{
\method{ggplot}{prop.huber.k}(fit, kstart = 0.5, kend = 3, kstep = 0.1, add.qvals = FALSE)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{kstart}{numeric, the first k estimated and shown in the plot.}

\item{kend}{numeric, the last k estimated and shown in the plot.}

\item{kstep}{numeric, controls the steps in which k is increased between kstart and kend.}

\item{add.qvals}{numeric, containing additional m-quantiles to be shown in plot.}
}
\description{
This plot shows the effect of k on the amount of observation being classified as outliers and therefore downweighted.
}
\examples{
library(mq1)

df <- simulate_data( n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "prop.huber.k")
}
