% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggplot.prop.res}
\alias{ggplot.prop.res}
\title{Proportion of Residuals smaller than 0 over the Fitted Values}
\usage{
\method{ggplot}{prop.res}(fit, add.qvals = FALSE)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}

\item{add.qvals}{numeric, containing additional m-quantiles to be shown in plot.}
}
\description{
This plot shows the proportion of residuals smaller than 0 over the fitted values.
This can give an indication of model fit and make m-quantile regression comparable with quantile regression.
}
\examples{
library(mq1)

df <- simulate_data(n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                  measurement.error = 0.5)

fit <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
plot(fit, plottype = "prop.res")

}
