% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmqm.R
\name{glmqm.fit}
\alias{glmqm.fit}
\title{Fitting generalized m-quantile models}
\usage{
glmqm.fit(y, X, family, weights, offset, q, control, glmrob.contr, ...)
}
\arguments{
\item{y}{}

\item{X}{}

\item{family}{}

\item{weights}{}

\item{offset}{}

\item{q}{}

\item{control}{}

\item{glmrob.contr}{}

\item{...}{}
}
\description{
Fitting generalized m-quantile models
}
