% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmqm.R
\name{glmqm}
\alias{glmqm}
\title{Generalized linear m-quantile model}
\usage{
glmqm(formula, data, subset, weights = NULL, weights.x = NULL,
  weights.var = NULL, na.action, family = gaussian, contrasts = NULL,
  tau = 0.5, start.fit, offset, control = glmqmControl(),
  rlm.control = rlmControl(), glmrob.control = glmrobControl(), ...)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{subset}{}

\item{weights}{}

\item{weights.x}{}

\item{weights.var}{}

\item{na.action}{}

\item{family}{}

\item{contrasts}{}

\item{tau}{}

\item{start.fit}{}

\item{offset}{}

\item{control}{}

\item{rlm.control}{}

\item{glmrob.control}{}

\item{...}{}
}
\description{
Generalized linear m-quantile model
}
