% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{glmqmControl}
\alias{glmqmControl}
\title{Controls for generalized linear m-quantile model}
\usage{
glmqmControl(maxit = 20, acc = 1e-04, k = 1.345, psi = psi.huber,
  sparse = FALSE, eps = .Machine$double.eps^(2/3), lm = "rlm")
}
\arguments{
\item{maxit}{}

\item{acc}{}

\item{k}{}

\item{psi}{}

\item{sparse}{}

\item{eps}{}

\item{lm}{}
}
\description{
Controls for generalized linear m-quantile model
}
