% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{glmrobControl}
\alias{glmrobControl}
\title{Controls for robust generalized linear model (robustbase::glmrob)}
\usage{
glmrobControl(method = "Mqle", method.control = glmrobMqle.control(),
  weights.on.x = "none", model = FALSE, x = FALSE, y = FALSE,
  trace.lev = 0, start = NULL)
}
\arguments{
\item{method}{}

\item{method.control}{}

\item{weights.on.x}{}

\item{model}{}

\item{x}{}

\item{y}{}

\item{trace.lev}{}

\item{start}{}
}
\description{
Controls for robust generalized linear model (robustbase::glmrob)
}
