% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{huberised}
\alias{huberised}
\alias{huberised.mquantfit}
\title{Show Huberised Residuals}
\usage{
huberised(x)

\method{huberised}{mquantfit}(x)
}
\arguments{
\item{x}{a mquantfit object, produced by calling the mquantreg-function.}
}
\value{
a dataframe-object with booleans indicating if the observation has been Huberised.
}
\description{
The huberised-method returns a flag indicating if an observation has been down-weighted by the alghorithm. .
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit  <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
huberised(fit)

}
