% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmqm.R
\name{lmqm.fit}
\alias{lmqm.fit}
\title{Fitting linear m-quantile model}
\usage{
lmqm.fit(y, X, weights, weights.x, weights.var, q, control, rlm.control, ...)
}
\arguments{
\item{y}{}

\item{X}{}

\item{weights}{}

\item{weights.x}{}

\item{weights.var}{}

\item{q}{}

\item{control}{}

\item{rlm.control}{}

\item{...}{}
}
\description{
Fitting linear m-quantile model
}
