% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqpoisson.R
\name{mqpoisson}
\alias{mqpoisson}
\title{M-Quantile Estimation for Count Data with Poisson Distribution}
\usage{
mqpoisson(
  x,
  y,
  q = 0.5,
  k = 1.6,
  maxit,
  acc = 1e-04,
  offset = NULL,
  case.weights = NULL,
  weights.x = FALSE,
  var.weights = NULL,
  qgrid = c(0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, seq(from = 0.45, to = 0.55, by =
    0.005), 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9),
  compute.qscores = FALSE
)
}
\arguments{
\item{x}{matrix, containing independent variables.}

\item{y}{vector, containing dependent variable.}

\item{q}{number strictly between 0 and 1, which specifies the m-quantile to be estimated.
Can also be a vector of m-quantiles.}

\item{k}{a number greater than 0. k is the parameter for the huber psi-function and
defaults to 1.6 for standalone use.}

\item{maxit}{a integer, which defines the maximum number of iterations
before the estimation process is halted. If the maximum number of iterations is
reached a warning message will be prompted.
If the maximal number of iteration is reached a warning message will be prompted.}

\item{acc}{defines convergence criteria.}

\item{offset}{a numeric vector of length equal to the number of observations.
This can be used to specify an a priori known component to be
included in the linear predictor during fitting. Defaults to \code{NULL}.}

\item{case.weights}{a numeric vector of observation weights;
if supplied, the algorithm fits to minimize the sum of the weights multiplied into
the absolute residuals.The length of case.weights must be the same as the number of
observations. The weights must be nonnegative and it is recommended that they be
strictly positive.Defaults to \code{NULL}.}

\item{weights.x}{boolean, for the estimation of robust glm to
obtain initial coefficiants in the discrete models.Defaults to \code{FALSE}.}

\item{var.weights}{dataframe, if supplied the residuals are scaled.}

\item{qgrid}{vector of quantiles, which are to be used for qscore calculations.}

\item{compute.qscores}{boolean, controls if qscores are to be computed.Defaults to
\code{FALSE}.}
}
\value{
See \code{\link{mquantreg}} for details.
\code{summary()}, \code{print()}, \code{fitted()} and \code{predict()}-methods are avaiable.
}
\description{
The function implements m-quantile estimation for count data as proposed by Tzavidis et al. (2015).
The algorithm is based on robust quasi-likelihood estimation by Cantoni & Ronchetti (2001).
The function is called by the mquantreg-function, when method "poisson" is selected.
Important: Standalone use is possible, but not advised.
}
\examples{

library(mq1)

df <- simulate_data(n = 100,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "count.poisson",
                   measurement.error = 0.05)

fit <- mquantreg(formula = Y ~ x1 + x2,data = df, q  = 0.5, method = "poisson")
print(fit)
}
\references{
\itemize{
\item Cantoni, E., & Ronchetti, E. (2001). Robust inference for generalized linear models. Journal of the American Statistical Association, 96(455), 1022-1030.
\item Tzavidis, N., Ranalli, M. G., Salvati, N., Dreassi, E., & Chambers, R. (2015). Robust small area prediction for counts. Statistical methods in medical research, 24(3), 373-395.
}
}
