% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{mquantreg}
\alias{mquantreg}
\title{M-Quantile Regression}
\usage{
mquantreg(
  formula,
  data,
  q = 0.5,
  k = 1.345,
  method = "continuous",
  scale.estimator = "Mad",
  na.action = na.fail,
  maxit = 1000,
  offset = NULL,
  case.weights = NULL,
  var.weights = NULL,
  weights.x = FALSE,
  acc = 1e-04,
  compute.qscores = FALSE,
  qgrid = c(0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, seq(from = 0.45, to = 0.55, by =
    0.005), 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9)
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a \code{~} operator
and the covariates, separated by \code{+} operators on the right hand side.}

\item{data}{a data.frame, which contains the variables named in the formula
and in the subset and the weights arguments.}

\item{q}{number strictly between 0 and 1, which specifies the m-quantile to be estimated.
Can also be a vector of m-quantiles.}

\item{k}{a number greater than 0. k is the parameter for the huber psi-function and
defaults to 1.345.}

\item{method}{string, which specifies the model to estimate.
The available models are: \cr
\itemize{
\item \code{"continuous"} - Continuous M-Quantile Regression. See \link{mqcont}.
\item \code{"binom"} - Binary M-Quantile Regession. See \link{mqbinom}.
\item \code{"poisson"} - Poisson M-Quantile Regession for Count Data. See \link{mqpoisson}.
\item \code{"quasipoisson"} - Quasipoisson M-Quantile Regession for Count Data. See \link{mqquasipoisson}.
\item \code{"nb"} - Negative Binomial  M-Quantile Regression for Count Data. See \link{mqnb}.
}}

\item{scale.estimator}{string, defines which function is used to estimate the
scale parameter used for peliminary estimation in the continuous regression.
The following scale estimators are available: \cr
\itemize{
\item \code{"Mad"} - Mean Absolute Deviation (default)
\item \code{"cMad"} - Corrected Mean Absolute Deviation
}}

\item{na.action}{defaults to \code{na.fail}.}

\item{maxit}{a integer, which defines the maximum number of iterations
before the estimation process is halted. If the maximum number of iterations is
reached a warning message will be prompted.}

\item{offset}{a numeric vector of length equal to the number of observations.
This can be used to specify an a priori known component to be
included in the linear predictor during fitting. Defaults to \code{NULL}.}

\item{case.weights}{a numeric vector of observation weights;
if supplied, the algorithm fits to minimize the sum of the weights multiplied into
the absolute residuals.The length of case.weights must be the same as the number of
observations. The weights must be nonnegative and it is recommended that they be
strictly positive.Defaults to \code{NULL}.}

\item{var.weights}{dataframe, if supplied the residuals are scaled. Has to be of
same length as number of observations.}

\item{weights.x}{boolean, for the estimation of robust glm to
obtain initial coefficiants in the discrete models.Defaults to \code{FALSE}.}

\item{acc}{defines convergence criteria.}

\item{compute.qscores}{boolean, controls if qscores are to be computed.Defaults to
\code{FALSE}.}

\item{qgrid}{vector of quantiles, which are to be used for qscore calculations.}
}
\value{
mquantreg returns an object of class "mquantfit". The class has a number of avaiable methods:\cr
\code{summary()}, \code{print()}, \code{fitted()}, \code{qscore()}, \code{huberise()} and \code{r2pseudo() }.
An object of class "mquantfit" is a list containing at least the following components:
\itemize{
\item coefficients - a named vector of coefficients.
\item residuals - the residuals, that is response minus fitted values.
\item fitted.values - the fitted mean values.
\item q.scores - a vector with the q-score of each observation
\item var.beta - variance of the estimators.
\item scale - estimated scale parameter.
}
Additional components are available, including all arguments passed when calling the mquantreg-function.
}
\description{
The mquantreg-function implements m-quantile regression for multiple types of
dependent data.It serves as an interface encompassing all types of m-quantile
regression implemented in this package. It is recommended to always use the mquantreg-
function, although it is possible to call the models from their respective functions.
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",

                  measurement.error = 0.5)
fit <- mquantreg(formula = Y ~ x1 + x2, data = df, q  = 0.05, method = "continuous")
summary(fit)
}
\references{
\itemize{
\item Breckling, J., & Chambers, R. (1988). M-quantiles. Biometrika, 75(4), 761-771.
\item Bianchi, A., Fabrizi, E., Salvati, N., & Tzavidis, N. (2015). M-quantile Regression:
Diagnostics and the parametric representation of the model.Book of Abstracts SIS - CLADAG, 303-306
}
}
