% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_testing.R
\name{mqwald}
\alias{mqwald}
\title{Wald test for M-quantile regression coefficients and linear hypotheses}
\usage{
mqwald(object, id.beta = "all", q = "all", R = NULL, H0 = NULL, each = FALSE)
}
\arguments{
\item{object}{an object of class "mquantfit", usually, a result of a call to \link{mquantreg}}

\item{id.beta}{identifier for the coefficient(s) to be jointly tested. Expects vector of positions
in coefficients of \code{object}. Defaults to "all", which does not include the intercept.}

\item{q}{identifies the m-quantile(s) for which coefficients should be tested.
Expects vector of m-quantiles e.g. \code{c(0.3, 0.5, 0.7)}.Defaults to "all".}

\item{R}{an optional matrix conformable to the coeffiecients from \code{object} in a way that
\code{R \%*\% beta.q} gives linear combinations of coefficients to be tested. Defaults to NULL.}

\item{H0}{a vector specifying the null hypotheses.}

\item{each}{gives the user the possibility to test each coefficient individually. If set to \code{TRUE},
all chosen coefficients in id.beta will be tested.Defaults to FALSE.}

\item{coefficents}{the estimated coeffients}

\item{varCov}{the variance-covariance matrix of the coefficients}

\item{statistic}{the test statistic}

\item{p.values}{the p-values for the Wald tests}
}
\value{
the function mqwald returns the the result of
}
\description{
This function allows to test the Null hypothesis of regression coefficients being zero.
It only works for coefficients estimated with a likelihood approach for a continuous dependent variable.
}
\examples{
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                  measurement.error = 0.05)

fit <- mquantreg(formula = Y ~ x1 + x2, data = df, q  = 0.5, method = "continuous")

hyp_test <- mqwald(fit)
print(hyp_test)

}
\references{
Bianchi, A., Fabrizi, E., Salvati, N., & Tzavidis, N. (2015).
M-quantile Regression: Diagnostics and the parametric representation of the model.
Book of Abstracts SIS - CLADAG, 303-306
Bianchi, A., & Salvati, N. (2015). Asymptotic properties and variance estimators
of the M-quantile regression coefficients estimators. Commun.Stat. Theory., 44, 2016–2429
}
