% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{print.mquantfit}
\alias{print.mquantfit}
\title{Print M-Quantile Estimates}
\usage{
\method{print}{mquantfit}(x, ...)
}
\arguments{
\item{x}{a mquantfit object, produced by calling the mquantreg-function.}

\item{...}{further arguments passed to or from other methods.
\code{summary()}, \code{print()}, \code{fit()} and \code{predict()}-methods are avaiable.}
}
\description{
The Print-method returns the coefficients for all estimated m-quantile in one table.
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
print(fit)
}
