% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{qscores}
\alias{qscores}
\alias{qscores.mquantfit}
\title{Compute Q-Scores}
\usage{
qscores(x, ...)

\method{qscores}{mquantfit}(
  x,
  qgrid = c(0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, seq(from = 0.45, to = 0.55, by =
    0.005), 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9)
)
}
\arguments{
\item{x}{a mquantfit object, produced by calling the mquantreg-function.}
}
\value{
a dataframe-object with the qscores of all observations.
}
\description{
The qscores()-method returns the qscores compute via inverse mquantile estimation. .
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit <-  mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
qscores(fit)

}
