% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{r2pseudo}
\alias{r2pseudo}
\alias{r2pseudo.mquantfit}
\title{Pseudo R-Squared for Continuous M-Quantile Estimation}
\usage{
r2pseudo(fit)

\method{r2pseudo}{mquantfit}(fit)
}
\arguments{
\item{fit}{a mquantfit object, produced by calling the mquantreg-function.}
}
\value{
numeric, the pseudo r-square value.
}
\description{
Computes the pseudo R-Squared for continuous M-Quantile Models with MAD scale estimator. The procedure was proposed by Bianchi (2015).
Bianchi, A., Fabrizi, E., Salvati, N. and Tzavidis, N. (eds) (2015). M-quantile re-
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
r2pseudo(fit)

}
