% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{rlmControl}
\alias{rlmControl}
\title{Controls for robust linear model (MASS::rlm)}
\usage{
rlmControl(init = "ls", scale.est = "MAD", method = "M",
  test.vec = "resid")
}
\arguments{
\item{init}{}

\item{scale.est}{}

\item{method}{}

\item{test.vec}{}
}
\description{
Controls for robust linear model (MASS::rlm)
}
