% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{stepmq}
\alias{stepmq}
\title{Backward stepwise variable selection using qscores.}
\usage{
stepmq(object, scope, threshold = 0.95, trace = 1, keep = NULL, steps = 1000)
}
\arguments{
\item{object}{object containing regression model fitted with mquantreg}

\item{scope}{defines the range of models examined in the stepwise search. This should be either a single formula,
or a list containing the component lower, a formula. See the details for how to specify the formula and how the argument is used.}

\item{threshold}{number strictly between 0 and 1. variables get dropped, if correlation of qscores
with model from step before exceeds the threshold.}

\item{trace}{if positive, information is printed during the running of stepmq.}

\item{keep}{a filter function whose input is a fitted model object and the associated qscores, and whose output is arbitrary.
Typically keep will select a subset of the components of the object and return them. The default is not to keep anything.}

\item{steps}{the maximum number of steps to be considered. The default is 1000 (essentially as many as required). It is
typically used to stop the process early.}
}
\value{
See \code{\link{mquantreg}} for details.
}
\description{
The function implements a backward variable selection method based on the correlation coefficient
of the models respective qscores.
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit <- mquantreg(formula = Y ~ x1 + x2, data = df, q  = 0.5, method = "continuous")
stepmq(fit)
}
\references{
Dawber, J. (2017). Advances in M-quantile estimation, University of Wollongong Thesis Collections
}
