% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mquantreg.R
\name{summary.mquantfit}
\alias{summary.mquantfit}
\title{Regression Output for Estimated M-Quantile Model}
\usage{
\method{summary}{mquantfit}(object, ...)
}
\arguments{
\item{object}{a mquantfit object, produced by calling the mquantreg-function.}

\item{...}{additional arguments affecting the summary produced
\code{summary()}, \code{print()}, \code{fit()} and \code{predict()}-methods are avaiable.}
}
\description{
The Summary()-method prints the regression output. The regression output includes
the coefficients and standard errors. Furthermore, information about the residuals and
the Huberisation are printed. For continuous estimation also pseudo R-square is computed.
The regression output for each m-quantile is returned seperately.
}
\examples{
library(mq1)
df <- simulate_data(n = 1000,
                  real.betas = c(0.1, 0.3, 0.1 ),
                  response.type = "continuous.normal",
                   measurement.error = 0.5)
fit <- mquantreg(formula = "Y ~ x1 + x2",data = df, q  = 0.5, method = "continuous")
summary(fit)
}
